pyproject.toml
[tool.isort]
profile = "black"
extend_skip_glob = ["alembic/*"]

[tool.poetry]
name = "Referrals API"
version = "0.1.0"
description = "REST API for referrals"
authors = ["70811772+SweetBubaleXXX@users.noreply.github.com"]

[tool.poetry.dependencies]
alembic = "^1.13.1"
dependency-injector = "^4.41.0"
fastapi = "^0.109.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
uvicorn = "^0.27.0.post1"
email-validator = "^2.1.0.post1"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.19.0"
asgi-lifespan = "^2.1.0"
httpx = "^0.26.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.4"
factory-boy = "^3.3.0"


[tool.poetry.group.prod.dependencies]
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
